requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
flasgger==0.9.5
Flask==2.1.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
mistune==2.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
